requirements.txt
Django==4.2.7
djangorestframework==3.15.1
django-cors-headers==4.0.0
gunicorn==21.2.0
